Splitting up provisioner into the requiste parts
Now we have an AccessPoint and Directory provisioner they should live in different files with their own tests.
We also need an OsClient to stop us having to run the tests as root which could cause issues in CI/CD.
