+++ pkg/driver/controller.go
@@ -105,7 +105,7 @@ func (d *Driver) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest)
 		return nil, status.Error(codes.InvalidArgument, "Volume ID not provided")
 	}
 
-	_, _, accessPointId, err := parseVolumeId(volId)
+	_, subpath, accessPointId, err := parseVolumeId(volId)
 	if err != nil {
 		//Returning success for an invalid volume ID. See here - https://github.com/kubernetes-csi/csi-test/blame/5deb83d58fea909b2895731d43e32400380aae3c/pkg/sanity/controller.go#L733
 		klog.V(5).Infof("DeleteVolume: Failed to parse volumeID: %v, err: %v, returning success", volId, err)
@@ -118,8 +118,13 @@ func (d *Driver) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest)
 		if err != nil {
 			return nil, status.Errorf(codes.Internal, "Failed to Delete volume %v: %v", volId, err)
 		}
+	} else if subpath != "" {
+		err := d.provisioners[DirectoryMode].Delete(ctx, req)
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "Failed to Delete volume %v: %v", volId, err)
+		}
 	} else {
-		return nil, status.Errorf(codes.NotFound, "Failed to find access point for volume: %v", volId)
+		return nil, status.Errorf(codes.NotFound, "Failed to find identifying information for volume: %v", volId)
 	}
 
 	return &csi.DeleteVolumeResponse{}, nil
